chore(tvix): dependency diet for tvix (part 1/?)
About a 7% dependency count reduction by dropping a lot of the tracing stuff
and, finally, the bigtable client.

Bigtable really was the boiling point for me: Bigtable doesn't even really exist
anymore, especially the Cloud one, which is emulated on Spanner! And what is the
point of this anyways?

The new direction for Tvix will include removing more and more of this stuff,
and replacing it with simpler (and less featureful, but hopefully useful!)
components.

Change-Id: I8be266cacc40a68bfc4eec883ed2f2a8069fd26b
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13375
Reviewed-by: sterni <sternenseemann@systemli.org>
Autosubmit: tazjin <tazjin@tvl.su>
Tested-by: BuildkiteCI