refactor(tvix/glue): move to simulated store implementation
As a first step towards removing the current store implementation this switches
over to running `EvalIO` and path imports through the new simulated store.

It's quite obvious that the current interfaces are not correct, and should not
be considered stable: There's hashing going on all over the place, values are
being recomputed, code duplication and so on.

This will need to be untangled, but step-by-step.

Note: This does not yet remove any dependencies on the old store implementation
in glue.

Change-Id: Ia7ae23afc7f68a3bf6c4c19c87e32a2430995291
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13404
Reviewed-by: sterni <sternenseemann@systemli.org>
Tested-by: BuildkiteCI