refactor(tvix/simstore): use walkdir iterator for path imports
Uses the `walkdir` crate used in glue code when filtering sources to perform
path imports in the simstore. The main difference is that walkdir provides a
flat list of entries (with depth metadata), which makes it compatible with the
filtering interface.

Using this directly for import_path means that we don't have to write similar
code twice to make filtered / non-filtered ingestion work properly.

Change-Id: Ie1c2d72865fc638e2d36723552202ab15e70a99e
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13446
Reviewed-by: sterni <sternenseemann@systemli.org>
Autosubmit: tazjin <tazjin@tvl.su>
Tested-by: BuildkiteCI