chore(tvix/build): delete project
This project is no longer relevant to Tvix, but if anyone needs this, go to
Snix, where it continues to live.

When we reimplement building based on our simpler store we might still take some
bits and pieces from here, but it seems like most of the important logic we'd
need to redo is already handled in the build bits that are left in the glue
project.

Change-Id: Ic27cb8c11edebd96175f36801f1b2eed142bcf5a
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13454
Reviewed-by: sterni <sternenseemann@systemli.org>
Tested-by: BuildkiteCI