refactor(tvix/eval): simplify attribute set representations
In our current setup, the multiple representations have no benefit - rather the
opposite, the construction cost is not worth it.

I benchmarked some versions closer to C++ style `std::flat_map`-ish
implementations (like in cppnix), but these do not help.

Lets throw this all out, and use just the one sane representation instead.

This actually improves performance for several reasons; one of the being that
the transition to bytestrings missed some cases that lead to unintended copies,
and just the general reduced overhead of the variant switching and extra
surrounding logic.

Change-Id: I28d45315c04cdd9e98612c03c1843db19d7c71ef
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13460
Tested-by: BuildkiteCI
Reviewed-by: sterni <sternenseemann@systemli.org>