refactor(tvix/eval): avoid some work when comparing attribute sets
The sorting requirement was sticking out as a hotspot, and seems to only have
been introduced at some point because it differs from cppnix.

I'm more of the "doesn't affect nixpkgs eval" tribe at the moment, and think we
can drop this.

The test that existed for this actually was introduced very recently (about two
weeks ago), and doesn't seem to have been tied to a real-world problem.

Change-Id: Ic5e62cd8ce5597a76982a5ea5bf3d9bf63e98b01
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13461
Tested-by: BuildkiteCI
Reviewed-by: sterni <sternenseemann@systemli.org>