refactor(tvix/eval): vu128 -> integer-encoding
This implementation has a more flexible encoding/decoding interface,
which allows us to avoid some copies when deserialising that were
showing up in hotspots.

The issue occured because we often have a varint near the end of a
chunk (in more than 50% of cases with random nixpkgs packages!), and the
copying that was required for that incured a penalty.

Change-Id: I4f0af4987837655c0a2c0b2e804258bc30c5349d
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13463
Tested-by: BuildkiteCI
Reviewed-by: sterni <sternenseemann@systemli.org>