chore(tvix/docs): clean up documentation of removed components
This change is a pretty low effort cleanup. The docs thingy would
benefit from some love, whipping it into a change that actually makes
sense. Currently, it is more or less random heap of markdown files that
doesn't, on the whole, make sense as an mdbook (which we also don't
serve anywhere).

Change-Id: I5193d67cc9aa52f25e8beb9ea9f3c8caf4b0a8f0
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13521
Autosubmit: sterni <sternenseemann@systemli.org>
Reviewed-by: tazjin <tazjin@tvl.su>
Tested-by: BuildkiteCI