feat(tvix/simstore): add nyi fetcher builtins to simstore
We need this _now_ for a very annoying reason: Someone decided it was a
good use of everyone's time to refactor Nixpkgs so that no builtin is
used with a redundant builtins.… prefix. Tvix dutifully checks whether
an accessed local is actually in scope if possible so compiling files
that mention fetchGit in some code path that's never taken (which
Nixpkgs has to support some cargo features) now fails.

Change-Id: I67df28088cba3dec01eed505176422c76849e127
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13826
Reviewed-by: tazjin <tazjin@tvl.su>
Tested-by: BuildkiteCI
Autosubmit: sterni <sternenseemann@systemli.org>