fix(tvix/simstore): readable paths only depend on store path existence
We only allow reads to already existing store paths at the moment.
We have been checking for the path the IO backend is asked to read from
instead of the (sometimes parent) store path to determine that, however.

Changing this to check for the store paths existence only allows for:

- Correct error messages when an expression tries to read from a missing
  path in the store: We can't realise subpaths of store paths, they're
  just missing.
- Using pathExists in subpaths of storePaths. The only thing we can't
  do is check for store path existence which can be done in regular Nix.
  This would require taking string context into account. For Nixpkgs,
  this is not relevant.

One assumption we make (which is alright) is that there are no partially
realised store paths.

Change-Id: I9dcf0684c4b9f20459fa6183bb1790c7bd94b567
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13896
Reviewed-by: tazjin <tazjin@tvl.su>
Tested-by: BuildkiteCI
Autosubmit: sterni <sternenseemann@systemli.org>